requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
pydub==0.25.1
PyPDF2==3.0.1
python_docx==1.1.2
SpeechRecognition==3.10.3
zhipuai==2.0.1.20240423.1
